+++ minesweeper_v1/main_utils.h
@@ -12,7 +12,7 @@ int printCenteredText(const std::string &text, const int &coordY);
 void printColoredTextWrapper(std::function<void()> func,
                              const int &backgroundColor, const int &textColor);
 int mod(const int &a, const int &b);
-std::string intToString(const int &number);
+std::string intToString(const long long &number);
 long long getTimeDiff(
     std::chrono::high_resolution_clock::time_point timePoint1,
     std::chrono::high_resolution_clock::time_point timePoint2);
